+++ packages/core/app/components/sections/ProductTiles/index.ts
