+++ packages/core/app/components/sections/ProductTiles/section.module.scss
