+++ packages/core/app/components/skeletons/ProductTilesSkeleton/ProductTileSkeleton/index.ts
