+++ packages/core/app/components/skeletons/ProductTilesSkeleton/ProductTilesSkeleton.tsx
@@ -1,6 +1,6 @@
 import type { PropsWithChildren } from 'react'
 
-import Tiles, { Tile } from 'src/components/ui/Tiles'
+import Tiles, { Tile } from 'app/components/ui/Tiles'
 
 import { ProductTileSkeleton } from '.'
